+++ packages/components/editor/src/styles.ts
@@ -6,7 +6,7 @@ export const editorVariants = tv({
   slots: {
     base: "group flex flex-col gap-1",
     editor:
-      "border-2 border-default-200 p-2 ui-group-hover:border-default-400 ui-group-focus:border-primary ui-group-disabled:cursor-not-allowed",
+      "border-2 border-default-200 p-2 uig-hover:border-default-400 uig-focus:border-primary uig-disabled:cursor-not-allowed",
     content:
       "max-h-[400px] min-h-[200px] overflow-y-auto border-t border-divider py-2",
     helperText: "text-xs text-foreground-400",