+++ packages/components/switch/src/styles.ts
@@ -7,14 +7,14 @@ export const switchVariants = tv({
     switch: [
       ...focusVisibleClasses,
       "flex cursor-pointer items-center rounded-full bg-default-200 p-1",
-      "ui-group-disabled:cursor-not-allowed ui-group-disabled:opacity-60",
+      "uig-disabled:cursor-not-allowed uig-disabled:opacity-60",
       "cursor-default ui-interactive:cursor-pointer ui-interactive:hover:opacity-95",
-      "justify-start ui-group-checked:justify-end",
+      "justify-start uig-checked:justify-end",
     ],
     thumb: "size-8 rounded-full bg-white shadow-md",
     helperText: "text-xs text-foreground-400",
     content: "grid grow gap-1",
-    label: "text-sm ui-group-interactive:cursor-pointer",
+    label: "text-sm uig-interactive:cursor-pointer",
     description: "text-xs text-foreground-500",
     mainWrapper: "flex items-center justify-between gap-2",
     switchWrapper: "",
@@ -22,22 +22,22 @@ export const switchVariants = tv({
   variants: {
     color: {
       default: {
-        switch: "ui-group-checked:bg-default",
+        switch: "uig-checked:bg-default",
       },
       primary: {
-        switch: "ui-group-checked:bg-primary",
+        switch: "uig-checked:bg-primary",
       },
       secondary: {
-        switch: "ui-group-checked:bg-secondary",
+        switch: "uig-checked:bg-secondary",
       },
       success: {
-        switch: "ui-group-checked:bg-success",
+        switch: "uig-checked:bg-success",
       },
       warning: {
-        switch: "ui-group-checked:bg-warning",
+        switch: "uig-checked:bg-warning",
       },
       danger: {
-        switch: "ui-group-checked:bg-danger",
+        switch: "uig-checked:bg-danger",
       },
     },
     size: {