+++ packages/components/select/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/select
 
+## 0.1.2
+
+### Patch Changes
+
+- fix group classes
+
 ## 0.1.1
 
 ### Patch Changes